- Refactored Attribute Scraping to be much more modular (allowing for easy expansion of attribute scraping in the future).
Instead of a default attribute scraping service that gets all types of attributes, there is a default delegate for scraping attributes.  This delegate can be overridden so that calling the new Scrape() scrapes whatever data is needed.

An example of how this works can be found in AttributeScrapingServiceIntegrationTests.cs in the WebScraping.Tests project.